Temp

+++ JDLC/c/include/connection/address.h
@@ -6,10 +6,10 @@
 struct address
 {
     char *host;
-    int16_t port;
+    uint32_t port;
     char *path;
 };
 
-struct address init_addr(const char *host, int16_t port, const char *path);
+struct address init_addr(const char *host, uint32_t port, const char *path);
 
 #endif
