+++ JDLC/c/src/connection/address.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <stdlib.h>
 
-struct address init_addr(const char *host, int16_t port, const char *path)
+struct address init_addr(const char *host, uint32_t port, const char *path)
 {
     struct address addr = {.host = (char *) malloc(strlen(host) * sizeof(char)),
                             .port = port,
